Create sep3_trybutton_timecount
There are 2 functions: firefly running a strip of 6 leds and fireflies running a strip of 3 leds.

Every time I push the button between 0-3 seconds fireflies function runs.
When I push it between 3-6 seconds firefly function runs.

In theory if I'd try this with the 39 leds strip it should work too. However, I've been trying out and after running it, one led in each array stays On and I'm not sure why.
For now, check this code that should work with the 3 leds you took.